requirements.txt
aiohttp==3.8.1
python-dotenv==0.20.0
requests==2.27.1
rich==12.2.0
trio==0.20.0
streamlink==3.2.0